// src/interfaces/professionals.interface.tsx
// Generated by https://quicktype.io

export interface IProfessionals {
    count: number;
    next: string;
    previous: string;
    results: Result[];
}

export interface Result {
    id: number;
    profile_image: string;
    first_name: string;
    last_name: string;
    email: string;
    is_active: boolean;
}

// Generated by https://quicktype.io

export interface IProfesssional {
    id: number;
    profile_image: string;
    first_name: string;
    last_name: string;
    email: string;
    is_active: boolean;
}

// Generated by https://quicktype.io

export interface IProfessionalLenguage {
    id: number;
    professional: IProfesssional;
    language: ILanguage;
    created_at: string;
    modified_at: string;
}

export interface ILanguage {
    id: number;
    name: string;
    code: string;
    is_active: boolean;
}

export interface FormCreateLanguage {
    language_id: number;
    professional_id: number;
}